Created separate files for filter related files.
